Cris 21/11/2022 V1
-> SQL <-
Creation of new assets in the list to have more data to work with.
Creation of a stored procedure that will return a list of main KPIs related to the assets status.

-> Web Page <-
Code to call a stored procedure that returns the cursor created in the function to be able to complete the main screen KPIs.