Cris 22/11/2022 V2
-> SQL
Modification of the procedure FUNC_ESTATUS_ACTIVOS in the select of the cursor, I included the two pending indicators required to complete the main screen indicators.

In addition, I add some validations information to be able to perform the select and the calculations

-> Web Page
I complete the main screen (My Dashboard) with the visualization and the indicators, no required to perform any addition modification to the main.

*Pending to validate if that calculates the risk level.