Cris 23/11/2022 V1
--> SQL
Creation of the table ASIENTO_LINEA in addition I added some inserts to the tables ASIENTO and ASIENTO_LINEA to be able to create some views.

I created the view RESUMEN_ASIENTOS used to create a table with all the journals posted in the Frond End. In addition I create the view ECUACION_CONTABLE to also complete journals information.

-> Front End
I include all the code to be able to consume the view RESUMEN_ASIENTOS and present the information in a table format.

Pending to add to the code the call to the view ECUACION_CONTABLE and show it in the screen.